requirements.txt
accelerate==0.22.0
asttokens==2.4.0
autoflake==2.2.1
backcall==0.2.0
black==23.7.0
boltons==23.0.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
cmake==3.27.4.1
comm==0.1.4
contourpy==1.1.0
cycler==0.11.0
debugpy==1.7.0
decorator==5.1.1
diffusers==0.20.2
exceptiongroup==1.1.3
executing==1.2.0
filelock==3.12.3
fonttools==4.42.1
fsspec==2023.9.0
huggingface-hub==0.16.4
idna==3.4
importlib-metadata==6.8.0
ipykernel==6.25.2
ipython==8.15.0
ipywidgets==8.1.0
jedi==0.19.0
Jinja2==3.1.2
jupyter_client==8.3.1
jupyter_core==5.3.1
jupyterlab-widgets==3.0.8
kiwisolver==1.4.5
lit==16.0.6
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mpmath==1.3.0
mypy-extensions==1.0.0
nest-asyncio==1.5.7
networkx==3.1
numpy==1.25.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
packaging==23.1
parso==0.8.3
pathspec==0.11.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.10.0
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyflakes==3.1.0
Pygments==2.16.1
pyparsing==3.0.9
python-dateutil==2.8.2
PyYAML==6.0.1
pyzmq==25.1.1
regex==2023.8.8
requests==2.31.0
safetensors==0.3.3
scipy==1.11.2
six==1.16.0
stack-data==0.6.2
sympy==1.12
tabulate==0.9.0
tokenizers==0.13.3
tomli==2.0.1
torch==2.0.1
torchaudio==2.0.2
torchsde==0.2.5
torchvision==0.15.2
tornado==6.3.3
tqdm==4.66.1
traitlets==5.9.0
trampoline==0.1.2
transformers==4.33.1
triton==2.0.0
typeguard==4.1.3
typing_extensions==4.7.1
urllib3==2.0.4
wcwidth==0.2.6
widgetsnbextension==4.0.8
zipp==3.16.2
